Update ppo.py to use Gymnasium
Updated for Gymnasium API. Unfortunately because seed is now passed is with env.reset() I was not able to get each environment in the vector environment to have a reproducible and distinct seed, but the code will now run at least. Also the video recorder wrapper does not seem to create any video files on my machine.